% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pythagorian_theorem.R
\name{Pythagorian_theorem}
\alias{Pythagorian_theorem}
\title{Pythagorian_theorem means give bounds}
\usage{
Pythagorian_theorem(x, y, z)
}
\arguments{
\item{x}{a vectror of numbers that will}

\item{s}{the lower}
}
\description{
This argument will take 3 arguments and return results
}
